Fixed a couple of warnings.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/branches/wip@3728 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/q3common.c
@@ -855,11 +855,11 @@ typedef struct {
 
 // field declarations
 #ifdef MSG_SHOWNET
-#	define PS_FIELD(n,b)	{ #n, ((int)&(((q3playerState_t *)0)->n)), b }
-#	define ES_FIELD(n,b)	{ #n, ((int)&(((q3entityState_t *)0)->n)), b }
+#	define PS_FIELD(n,b)	{ #n, ((size_t)&(((q3playerState_t *)0)->n)), b }
+#	define ES_FIELD(n,b)	{ #n, ((size_t)&(((q3entityState_t *)0)->n)), b }
 #else
-#	define PS_FIELD(n,b)	{ ((int)&(((q3playerState_t *)0)->n)), b }
-#	define ES_FIELD(n,b)	{ ((int)&(((q3entityState_t *)0)->n)), b }
+#	define PS_FIELD(n,b)	{ ((size_t)&(((q3playerState_t *)0)->n)), b }
+#	define ES_FIELD(n,b)	{ ((size_t)&(((q3entityState_t *)0)->n)), b }
 #endif
 
 // field data accessing
